Finish first draft of css for this project
I again forgot to commit early.
So this commit contains quite a lot of new content.
I changed the classes of the sections in main to ids.
I also finished the css file for this project.
For now I will leave this project,
but I might revisit it in the future.